removed tokenCache
